# apps/api-erp/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "AppController_getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Return API health status",
        "tags": [
          "health"
        ]
      }
    },
    "/auth/sign-up": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthContextDto"
                }
              }
            }
          }
        },
        "summary": "Create a user account and start a session",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/sign-in": {
      "post": {
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthContextDto"
                }
              }
            }
          }
        },
        "summary": "Sign in with mobile number or email and create a session cookie",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthContextDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get the current authenticated user context",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/session": {
      "get": {
        "operationId": "AuthController_getSession",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthContextDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get the current authenticated session context",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/context/campus": {
      "patch": {
        "operationId": "AuthController_selectCampus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchCampusBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthContextDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Select the active campus inside the current tenant",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/sign-out": {
      "post": {
        "operationId": "AuthController_signOut",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Delete the current session and clear the cookie",
        "tags": [
          "auth"
        ]
      }
    },
    "/onboarding/institutions": {
      "post": {
        "operationId": "OnboardingController_createInstitution",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInstitutionOnboardingBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthContextDto"
                }
              }
            }
          }
        },
        "summary": "Create an institution, default campus, and first admin account",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/institutions/{institutionId}/campuses": {
      "get": {
        "operationId": "CampusesController_listCampuses",
        "parameters": [
          {
            "name": "institutionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampusDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List campuses for an institution",
        "tags": [
          "campuses"
        ]
      },
      "post": {
        "operationId": "CampusesController_createCampus",
        "parameters": [
          {
            "name": "institutionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampusBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create a campus for an institution",
        "tags": [
          "campuses"
        ]
      }
    },
    "/institutions/{institutionId}/academic-years": {
      "get": {
        "operationId": "AcademicYearsController_listAcademicYears",
        "parameters": [
          {
            "name": "institutionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Academic years list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AcademicYearDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List academic years for an institution",
        "tags": [
          "academic-years"
        ]
      },
      "post": {
        "operationId": "AcademicYearsController_createAcademicYear",
        "parameters": [
          {
            "name": "institutionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAcademicYearBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Academic year created"
          }
        },
        "summary": "Create an academic year for an institution",
        "tags": [
          "academic-years"
        ]
      }
    },
    "/institutions/{institutionId}/academic-years/{academicYearId}/current": {
      "patch": {
        "operationId": "AcademicYearsController_setCurrentAcademicYear",
        "parameters": [
          {
            "name": "institutionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "academicYearId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Academic year set as current"
          }
        },
        "summary": "Set current academic year",
        "tags": [
          "academic-years"
        ]
      }
    },
    "/institutions/{institutionId}/academic-years/{academicYearId}/archive": {
      "patch": {
        "operationId": "AcademicYearsController_archiveAcademicYear",
        "parameters": [
          {
            "name": "institutionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "academicYearId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Academic year archived"
          }
        },
        "summary": "Archive an academic year",
        "tags": [
          "academic-years"
        ]
      }
    },
    "/institutions/{institutionId}/academic-years/{academicYearId}/restore": {
      "patch": {
        "operationId": "AcademicYearsController_restoreAcademicYear",
        "parameters": [
          {
            "name": "institutionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "academicYearId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Academic year restored"
          }
        },
        "summary": "Restore an academic year",
        "tags": [
          "academic-years"
        ]
      }
    },
    "/institutions": {
      "get": {
        "operationId": "InstitutionsController_listInstitutions",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "slug",
                "createdAt"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated institution list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListInstitutionsResultDto"
                }
              }
            }
          }
        },
        "summary": "List institutions",
        "tags": [
          "institutions"
        ]
      }
    },
    "/institutions/counts": {
      "get": {
        "operationId": "InstitutionsController_getInstitutionCounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Institution counters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstitutionCountsDto"
                }
              }
            }
          }
        },
        "summary": "Count institutions by status",
        "tags": [
          "institutions"
        ]
      }
    },
    "/institutions/{id}": {
      "get": {
        "operationId": "InstitutionsController_getInstitutionById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Institution details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstitutionDto"
                }
              }
            }
          },
          "404": {
            "description": "Institution not found"
          }
        },
        "summary": "Get institution by id",
        "tags": [
          "institutions"
        ]
      }
    },
    "/public/tenant-branding": {
      "get": {
        "operationId": "PublicController_getTenantBranding",
        "parameters": [
          {
            "name": "tenant",
            "required": false,
            "in": "query",
            "description": "Optional tenant slug override for localhost development.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantBrandingDto"
                }
              }
            }
          }
        },
        "summary": "Return tenant branding for the current subdomain or local dev override.",
        "tags": [
          "public"
        ]
      }
    }
  },
  "info": {
    "title": "Academic Platform ERP API",
    "description": "NestJS backend for ERP authentication, onboarding, tenant branding, and school operations.",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "health",
      "description": ""
    },
    {
      "name": "public",
      "description": ""
    },
    {
      "name": "auth",
      "description": ""
    },
    {
      "name": "onboarding",
      "description": ""
    },
    {
      "name": "campuses",
      "description": ""
    },
    {
      "name": "academic-years",
      "description": ""
    },
    {
      "name": "institutions",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "HealthResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
      "SignUpBodyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string"
          },
          "tenantSlug": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "mobile",
          "password"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "mobile",
          "email"
        ]
      },
      "AuthMembershipDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "organizationName": {
            "type": "string"
          },
          "organizationSlug": {
            "type": "string"
          },
          "memberType": {
            "type": "string",
            "enum": [
              "staff",
              "student",
              "guardian"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "suspended"
            ]
          },
          "primaryCampusId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "organizationId",
          "organizationName",
          "organizationSlug",
          "memberType",
          "status",
          "primaryCampusId"
        ]
      },
      "AuthOrganizationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "shortName": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "institutionType": {
            "type": "string",
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "faviconUrl": {
            "type": "string",
            "nullable": true
          },
          "primaryColor": {
            "type": "string"
          },
          "accentColor": {
            "type": "string"
          },
          "sidebarColor": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "shortName",
          "slug",
          "institutionType",
          "logoUrl",
          "faviconUrl",
          "primaryColor",
          "accentColor",
          "sidebarColor",
          "status"
        ]
      },
      "AuthCampusDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "isDefault": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          }
        },
        "required": [
          "id",
          "organizationId",
          "name",
          "slug",
          "code",
          "isDefault",
          "status"
        ]
      },
      "AuthContextDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthMembershipDto"
            }
          },
          "activeOrganization": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthOrganizationDto"
              }
            ]
          },
          "activeCampus": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthCampusDto"
              }
            ]
          },
          "campuses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthCampusDto"
            }
          }
        },
        "required": [
          "user",
          "expiresAt",
          "memberships",
          "activeOrganization",
          "activeCampus",
          "campuses"
        ]
      },
      "SignInBodyDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "tenantSlug": {
            "type": "string"
          }
        },
        "required": [
          "identifier",
          "password"
        ]
      },
      "SwitchCampusBodyDto": {
        "type": "object",
        "properties": {
          "campusId": {
            "type": "string"
          }
        },
        "required": [
          "campusId"
        ]
      },
      "CreateInstitutionOnboardingBodyDto": {
        "type": "object",
        "properties": {
          "institutionName": {
            "type": "string"
          },
          "institutionSlug": {
            "type": "string"
          },
          "institutionShortName": {
            "type": "string"
          },
          "campusName": {
            "type": "string"
          },
          "campusSlug": {
            "type": "string"
          },
          "adminName": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "institutionName",
          "institutionSlug",
          "campusName",
          "adminName",
          "mobile",
          "password"
        ]
      },
      "CampusDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "isDefault": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          }
        },
        "required": [
          "id",
          "organizationId",
          "name",
          "slug",
          "code",
          "isDefault",
          "status"
        ]
      },
      "CreateCampusBodyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "isDefault": {
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ]
      },
      "AcademicYearDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "isCurrent": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "archived"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "startDate",
          "endDate",
          "isCurrent",
          "status",
          "createdAt"
        ]
      },
      "CreateAcademicYearBodyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "makeCurrent": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "startDate",
          "endDate",
          "makeCurrent"
        ]
      },
      "InstitutionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "institutionType": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true,
            "enum": [
              "active",
              "suspended"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "institutionType",
          "status",
          "createdAt"
        ]
      },
      "ListInstitutionsResultDto": {
        "type": "object",
        "properties": {
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InstitutionDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "pageCount": {
            "type": "number"
          }
        },
        "required": [
          "rows",
          "total",
          "page",
          "pageSize",
          "pageCount"
        ]
      },
      "InstitutionCountsDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "active": {
            "type": "number"
          },
          "suspended": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "active",
          "suspended"
        ]
      },
      "TenantBrandingDto": {
        "type": "object",
        "properties": {
          "institutionName": {
            "type": "string"
          },
          "shortName": {
            "type": "string"
          },
          "tenantSlug": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "faviconUrl": {
            "type": "string",
            "nullable": true
          },
          "primaryColor": {
            "type": "string"
          },
          "accentColor": {
            "type": "string"
          },
          "sidebarColor": {
            "type": "string"
          }
        },
        "required": [
          "institutionName",
          "shortName",
          "tenantSlug",
          "logoUrl",
          "faviconUrl",
          "primaryColor",
          "accentColor",
          "sidebarColor"
        ]
      }
    }
  }
}